Unify JarWriter and com.yworks.yshrink.core.Writer
