+++ src/main/java/com/yworks/yshrink/core/Writer.java
@@ -5,30 +5,41 @@ import com.yworks.util.abstractjar.StreamProvider;
 import com.yworks.util.abstractjar.impl.DirectoryStreamProvider;
 import com.yworks.util.abstractjar.impl.DirectoryWrapper;
 import com.yworks.util.abstractjar.impl.JarFileWrapper;
+import com.yworks.util.abstractjar.impl.JarStreamProvider;
 import com.yworks.yguard.common.ResourcePolicy;
 import com.yworks.yguard.common.ShrinkBag;
 import com.yworks.yshrink.model.ClassDescriptor;
 import com.yworks.yshrink.model.Model;
-import com.yworks.util.abstractjar.impl.JarStreamProvider;
 import com.yworks.yshrink.util.Logger;
 import com.yworks.yshrink.util.Util;
+import com.yworks.yshrink.util.Version;
 import org.objectweb.asm.ClassReader;
 import org.objectweb.asm.ClassWriter;
 
+import java.io.BufferedOutputStream;
 import java.io.DataInputStream;
 import java.io.File;
+import java.io.FileOutputStream;
 import java.io.IOException;
 import java.security.MessageDigest;
 import java.security.NoSuchAlgorithmException;
 import java.text.NumberFormat;
 import java.util.HashSet;
 import java.util.Set;
+import java.util.jar.Attributes;
+import java.util.jar.JarEntry;
+import java.util.jar.JarOutputStream;
 import java.util.jar.Manifest;
 
 /**
  * @author Michael Schroeder, yWorks GmbH http://www.yworks.com
  */
 public class Writer {
+
+  private Set<String> directoriesWritten = new HashSet<String>();
+  private FileOutputStream fos;
+  private JarOutputStream jos;
+  private Manifest manifest;
 
   public static final String MANIFEST_FILENAME = "META-INF/MANIFEST.MF";
   private static final String SIGNATURE_FILE_PREFIX = "META-INF/";
@@ -63,6 +74,98 @@ public class Writer {
     return digests;
   }
 
+  private void addDigests( String entryName ) {
+    Attributes oldEntryAttributes = manifest.getAttributes(entryName);
+    Attributes newEntryAttributes = new Attributes(digests.length + 1);
+
+    if (null != oldEntryAttributes) {
+      Set<Object> keys = oldEntryAttributes.keySet();
+      for (Object key : keys) {
+        if (((Attributes.Name) key).toString().indexOf("Digest") == -1) {
+          newEntryAttributes.put(key, oldEntryAttributes.get(key));
+        }
+      }
+    }
+
+    StringBuffer digestsList = new StringBuffer();
+    for (int i = 0; i < digests.length; i++) {
+      MessageDigest digest = digests[i];
+
+      if (null != digest) {
+
+        String digestKey = digest.getAlgorithm() + "-Digest";
+        digestsList.append(digest.getAlgorithm());
+        if (i < digests.length - 1) {
+          digestsList.append(", ");
+        }
+
+        String digestVal = Util.toBase64(digest.digest());
+
+        newEntryAttributes.putValue(digestKey, digestVal);
+      }
+    }
+
+    newEntryAttributes.putValue("Digest-Algorithms", digestsList.toString());
+
+    this.manifest.getEntries().put(entryName, newEntryAttributes);
+  }
+
+  private void calcDigests( final byte[] data ) {
+    for (int i = digests.length - 1; i >= 0; i--) {
+      if (null != digests[i]) {
+        digests[i].reset();
+        digests[i].update(data);
+      }
+    }
+  }
+
+  private void addEntry( final String fileName, final byte[] data ) throws IOException {
+
+    JarEntry outEntry = new JarEntry(fileName);
+    addDirectory(fileName);
+    jos.putNextEntry(outEntry);
+    jos.write(data);
+    jos.closeEntry();
+
+    calcDigests(data);
+
+    addDigests(fileName);
+  }
+
+  private void addDirectory( final String fileName ) throws IOException {
+    int index = 0;
+    while ((index = fileName.indexOf("/", index + 1)) >= 0) {
+      String directory = fileName.substring(0, index + 1);
+      if (!directoriesWritten.contains(directory)) {
+        directoriesWritten.add(directory);
+        JarEntry directoryEntry = new JarEntry(directory);
+        jos.putNextEntry(directoryEntry);
+        jos.closeEntry();
+      }
+    }
+  }
+
+  private void finishManifest() throws IOException {
+    manifest.getMainAttributes().putValue("Created-by",
+                                          "yGuard Bytecode Obfuscator: Shrinker " + Version.getVersion());
+
+    addDirectory(Writer.MANIFEST_FILENAME);
+    jos.putNextEntry(new JarEntry(Writer.MANIFEST_FILENAME));
+    this.manifest.write(jos);
+    jos.closeEntry();
+  }
+
+  private void close() throws IOException {
+    finishManifest();
+    if (jos != null) {
+      jos.finish();
+      jos.close();
+    }
+    if (fos != null) {
+      fos.close();
+    }
+  }
+
   public void write( Model model, ShrinkBag bag ) throws IOException {
 
     File in = bag.getIn();
@@ -81,8 +184,9 @@ public class Writer {
 
     if ( !out.exists() ) out.createNewFile();
 
-    final Manifest newManifest = ( inJar.getManifest() != null) ? new Manifest( inJar.getManifest() ) : new Manifest();
-    final JarWriter writer = new JarWriter( out, newManifest, this );
+    manifest = ( inJar.getManifest() != null) ? new Manifest( inJar.getManifest() ) : new Manifest();
+    fos = new FileOutputStream(out);
+    jos = new JarOutputStream(new BufferedOutputStream(fos));
 
     int numClasses = 0;
     int numObsoleteClasses = 0;
@@ -127,9 +231,9 @@ public class Writer {
         numObsoleteFields += outputVisitor.getNumObsoleteFields();
 
         byte[] modifiedClass = cw.toByteArray();
-        writer.addEntry( entryName, modifiedClass );
+        addEntry( entryName, modifiedClass );
       } else {
-        newManifest.getEntries().remove( entryName );
+        manifest.getEntries().remove( entryName );
         numObsoleteClasses++;
         Logger.shrinkLog( "\t\t<class name=\"" + Util.toJavaClass( entryName ) + "\" />" );
       }
@@ -158,7 +262,7 @@ public class Writer {
                     ( resourcePolicy.equals( ResourcePolicy.AUTO ) &&
                         nonEmptyDirs.contains( jarStreamProvider.getCurrentDir() ) ) ) ) {
 
-          copyResource( entryName, jarStreamProvider, stream, writer );
+          copyResource( entryName, jarStreamProvider, stream );
         } else {
           numRemovedResources++;
           Logger.shrinkLog(
@@ -171,7 +275,7 @@ public class Writer {
 
     Logger.shrinkLog( "\t</removed-resources>" );
 
-    writer.close();
+    close();
 
     long outLength = out.length();
 
@@ -188,8 +292,7 @@ public class Writer {
     Logger.shrinkLog( "</inOutPair>" );
   }
 
-  private void copyResource( String entryName, StreamProvider jarStreamProvider, DataInputStream stream,
-                             JarWriter writer ) throws IOException {
+  private void copyResource( String entryName, StreamProvider jarStreamProvider, DataInputStream stream ) throws IOException {
 
     // don't copy manifest/signature files.
     if ( ! entryName.equals( MANIFEST_FILENAME )
@@ -199,7 +302,7 @@ public class Writer {
       if ( -1 != entrySize ) {
         byte[] data = new byte[ entrySize ];
         stream.readFully( data );
-        writer.addEntry( entryName, data );
+        addEntry( entryName, data );
       }
     }
   }
